requirements.txt
# Core Dependencies
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4

# Async and HTTP
aiofiles==23.2.1
requests==2.31.0
httpx==0.27.0
aiohttp==3.9.3

# Google API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
google-api-python-client==2.117.0

# LLM and Search
groq==0.4.0
google-search-results==2.4.2
beautifulsoup4==4.12.3

# HTML Processing
lxml[html_clean]==5.1.0

# Data Processing
python-dotenv==1.0.0
pydantic==2.6.3
chardet==5.2.0

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Utilities
tqdm==4.66.2
loguru==0.7.2
tenacity==8.2.3

# Type Checking
mypy==1.8.0
types-requests==2.31.0.20240218